Upgrade storybook to 7.5
