+++ lib/.eslintrc.json
@@ -4,20 +4,65 @@
     "browser": true,
     "window": true
   },
-  "extends": ["airbnb", "prettier", "plugin:jest/recommended"],
-  "plugins": ["jest", "react-hooks"],
+  "extends": [
+    "airbnb",
+    "prettier",
+    "plugin:jest/recommended",
+    "plugin:storybook/recommended",
+    "plugin:storybook/recommended"
+  ],
+  "plugins": [
+    "jest",
+    "react-hooks"
+  ],
   "rules": {
     "linebreak-style": 0,
-    "no-use-before-define": ["error", { "functions": false }],
-    "no-underscore-dangle": ["error", { "allow": ["_options", "_links"] }],
+    "no-use-before-define": [
+      "error",
+      {
+        "functions": false
+      }
+    ],
+    "no-underscore-dangle": [
+      "error",
+      {
+        "allow": [
+          "_options",
+          "_links"
+        ]
+      }
+    ],
     "react/jsx-one-expression-per-line": "allow",
     "no-nested-ternary": "off",
     "import/extensions": "allow",
-    "import/no-unresolved": ["error", { "ignore": ["react", "styled-components"] }],
-    "react/prop-types": [2, { "ignore": ["children"] }],
+    "import/no-unresolved": [
+      "error",
+      {
+        "ignore": [
+          "react",
+          "styled-components"
+        ]
+      }
+    ],
+    "react/prop-types": [
+      2,
+      {
+        "ignore": [
+          "children"
+        ]
+      }
+    ],
     "react/self-closing-comp": "allow",
     "react-hooks/rules-of-hooks": "error",
     "react-hooks/exhaustive-deps": "warn",
-    "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }]
+    "react/jsx-filename-extension": [
+      1,
+      {
+        "extensions": [
+          ".js",
+          ".jsx"
+        ]
+      }
+    ]
   }
 }
