--- lib/.storybook/preview.js
@@ -1,34 +1,0 @@
-import { HalstackProvider } from "../src/HalstackContext";
-import styled from "styled-components";
-import { INITIAL_VIEWPORTS } from "@storybook/addon-viewport";
-
-export const parameters = {
-  actions: { argTypesRegex: "^on[A-Z].*" },
-  controls: {
-    matchers: {
-      color: /(background|color)$/i,
-      date: /Date$/,
-    },
-  },
-  viewport: {
-    viewports: INITIAL_VIEWPORTS,
-  },
-};
-
-export const decorators = [
-  (Story) => (
-    <HalstackProvider>
-      <Container>
-        <Story />
-      </Container>
-    </HalstackProvider>
-  ),
-];
-
-const Container = styled.div`
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen", "Ubuntu", "Cantarell", "Fira Sans",
-    "Droid Sans", "Helvetica Neue", sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-`;
