+++ lib/.storybook/preview.tsx
@@ -1,0 +1,32 @@
+import { HalstackProvider } from "../src/HalstackContext";
+import styled from "styled-components";
+import type { Preview } from "@storybook/react";
+import React from "react";
+
+const preview: Preview = {
+  parameters: {
+    actions: { argTypesRegex: "^on[A-Z].*" },
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+  },
+  decorators: [
+    (Story) => (
+      <HalstackProvider>
+        <Container>
+          <Story />
+        </Container>
+      </HalstackProvider>
+    ),
+  ],
+};
+
+const Container = styled.div`
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen", "Ubuntu", "Cantarell", "Fira Sans",
+    "Droid Sans", "Helvetica Neue", sans-serif;
+`;
+
+export default preview;
